Make del-random more agressive, delete 1/5th of objects

+++ src/commands.rs
@@ -103,7 +103,7 @@ enum Commands {
         tub: Option<PathBuf>,
     },
 
-    #[command(about = "Super scary command to delete 1000 random objects")]
+    #[command(about = "Super scary command to delete 1/5 of the objects")]
     DelRandom {
         #[arg(short, long, value_name="DIR")]
         #[arg(help="Path of Tub control directory (defaults to CWD)")]
@@ -334,8 +334,11 @@ fn cmd_stats(tub: OptPath) -> io::Result<()>
 fn cmd_del_random(tub: OptPath) -> io::Result<()>
 {
     let mut tub = get_tub(tub)?;
-    for hash in tub.keys().iter().take(1000) {
-        tub.delete_object(hash)?;
+    let keys = tub.keys();
+    for (i, hash) in keys.iter().enumerate() {
+        if i % 5 == 0 {
+            tub.delete_object(hash)?;
+        }
     }
     eprintln!("{} objects in store", tub.len());
     Ok(())
